added optional arguments for cleaning up or not and some error handling

+++ batch_sorting_phase1.py
@@ -2,7 +2,15 @@
 import os, sys, time,glob
 
 
-def batch_sorting_phase1(incoming_folder,batch_folders,results_folder):
+def batch_sorting_phase1(incoming_folder,batch_folders,results_folder,sort_unknown=False,delete_closed=True):
+	#
+	# Optional arguments:
+	# sort_unknown: sort files that don't match the lists of open or closed batches into "unknown_batches" folder
+	#               For this option, the 'open_batches.txt' and 'closed_batches.txt'
+	#               need to be up to date otherwise files will be sorted incorrectly
+	#               Default is False
+	# delete_closed: delete files from closed batches
+	#               Default is True
 	print time.strftime("%Y/%m/%d %H:%M:%S") + " Starting batch_sorting_phase1.py\n"
 
 	open_batches=[]
@@ -23,40 +31,45 @@ def batch_sorting_phase1(incoming_folder,batch_folders,results_folder):
 	# Loop over zip files in incoming folder
 	for fpath in glob.glob(incoming_folder+'/*.zip'):
 		fname=os.path.basename(fpath)
-		# Split string up (assume has the format: hadam3p_eu_fs23_201412_1_d401_000000951_0_r1285010883_1.zip
-		str_split=fname.split('_')
-		fname_out=''
-		if str_split[-2][0]=='r':
-			batch=str_split[-5]
-			# recombine fname without random part
-			for i,part in enumerate(str_split):
-				if i != len(str_split)-2:
-					fname_out=fname_out+part+'_'
-			fname_out=fname_out[:-1] # remove final underscore
-		else:
-			# Old format without random string
-			batch=str_split[-4]
-			fname_out=fname
-		workunit_name=fname_out[:-len(str_split[-1])-1] # workunit name is fname without the '_1.zip'
-		# Check if workunit is from an open batch
-		if batch in open_batches:
-			batch_folder=results_folder+'/batch_'+batch
-			if not os.path.exists(batch_folder):
-				os.mkdir(batch_folder)
-			if not os.path.exists(batch_folder+'/in_progress/'):
-				os.mkdir(batch_folder+'/in_progress/')
-			if not os.path.exists(batch_folder+'/in_progress/'+workunit_name):
-				os.mkdir(batch_folder+'/in_progress/'+workunit_name)
-			os.rename(fpath,batch_folder+'/in_progress/'+workunit_name+'/'+fname_out)
-			print fname_out,batch
-		# Check if workunit is from a closed batch
-		elif batch in closed_batches:
-		# Delete this file
-			os.remove(fpath)
-			print 'deleting from closed batch',fname,batch
-		else:
-				os.rename(fpath,results_folder+'/unknown_batches/'+fname)
-				print fname,'unknown batch'
+		try:
+			# Split string up (assume has the format: hadam3p_eu_fs23_201412_1_d401_000000951_0_r1285010883_1.zip
+			str_split=fname.split('_')
+			fname_out=''
+			if str_split[-2][0]=='r':
+				batch=str_split[-5]
+				# recombine fname without random part
+				for i,part in enumerate(str_split):
+					if i != len(str_split)-2:
+						fname_out=fname_out+part+'_'
+				fname_out=fname_out[:-1] # remove final underscore
+			else:
+				# Old format without random string
+				batch=str_split[-4]
+				fname_out=fname
+			workunit_name=fname_out[:-len(str_split[-1])-1] # workunit name is fname without the '_1.zip'
+			# Check if workunit is from an open batch
+			if batch in open_batches:
+				batch_folder=results_folder+'/batch_'+batch
+				if not os.path.exists(batch_folder):
+					os.mkdir(batch_folder)
+				if not os.path.exists(batch_folder+'/in_progress/'):
+					os.mkdir(batch_folder+'/in_progress/')
+				if not os.path.exists(batch_folder+'/in_progress/'+workunit_name):
+					os.mkdir(batch_folder+'/in_progress/'+workunit_name)
+				os.rename(fpath,batch_folder+'/in_progress/'+workunit_name+'/'+fname_out)
+				print fname_out,batch
+			# Check if workunit is from a closed batch
+			elif batch in closed_batches:
+				if delete_closed:
+					# Delete this file
+					os.remove(fpath)
+					print 'deleting from closed batch',fname,batch
+			else:
+				if sort_unknown:
+					os.rename(fpath,results_folder+'/unknown_batches/'+fname)
+					print fname,'unknown batch'
+		except:
+			print "Error sorting file:",fname
 	print time.strftime("%Y/%m/%d %H:%M:%S") + " Finished batch_sorting_phase1.py"
 #####################
